Removed verbs form URI's. Simplified check if Order already contains Product. Added orderId to DuplicateOrderingOfProductException.
