+++ game_shop-orders/src/main/java/capgemini/gameshop/microservice/order/exception/DuplicateOrderingOfProductException.java
@@ -1,6 +1,7 @@
 package capgemini.gameshop.microservice.order.exception;
 
 public class DuplicateOrderingOfProductException extends RuntimeException{
-    public DuplicateOrderingOfProductException(Long productId) { super("Product with id: " + productId + " already selected.");
+    public DuplicateOrderingOfProductException(Long orderId ,Long productId) { super(
+            "Order with id: " + orderId + " already contains Product with id: " + productId + "!");
     }
 }
